refactor: extract team-sync-init.ts from daemon/main.ts
Move team sync client creation, node registration, backfill, and
outbox flush job into initTeamSync() factory. main.ts uses the
returned getter/setter/registerFlushJob interface.
